+++ drizzle/0018_special_colonel_america.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `song` ADD `frame_rate` real DEFAULT 25 NOT NULL;